fix(ssh-proxy): wire module into workspace, fix all hallucinated import paths (C1)
- Add ./apps/ssh-proxy to go.work (workspace member)
- Rewrite apps/ssh-proxy/go.mod: module renamed to
  github.com/mintkey/mintkey/services/ssh-proxy, go directive
  bumped to 1.26.0, replace directives removed, real workspace
  deps wired
- Create apps/ssh-proxy/internal/vault/client.go: per-service vault
  client matching proxy-plugin pattern; AuthSchemeSSHPrivateKey=11,
  TargetAddress/SSHUser on Credential, GetAgentByFingerprint /
  GetHostKeyFingerprint / StoreHostKeyFingerprint stubs returning
  ErrNotImplemented
- Replace all github.com/WeLikeCode/mintkey/internal/* import paths
  with github.com/mintkey/mintkey/packages/go/*
- Fix otelinit.Config/LoadConfig() usage: replace with OTelEndpoint
  string field in Config; add VaultIdentityID/VaultToken config
  fields (env: MINTKEY_VAULT_SSH_PROXY_IDENTITY_ID/TOKEN)
- Fix auditq.Queue.Enqueue API: was (ctx, tenantID, []byte), now
  takes auditq.Event directly
- Fix changes.Subscriber/Event: replace with changes.Client +
  WithEventHandler; define local revocation.Event type; update
  revocation_test.go to drop invalid changes.Event reference
- Fix bridge/pty.go uint32 vs int type mismatch
- Fix ulid.New("audit") / ulid.New("session") → trailing underscore
  required by ADR-0017.11
- Fix vault.NewClient(addr) → NewClient(addr, identityID, token)
- Fix hostkey.go GetHostKeyFingerprint/StoreHostKeyFingerprint
  argument count to match new stub signatures

+++ apps/ssh-proxy/internal/bridge/pty.go
@@ -39,11 +39,11 @@ func ParsePTYRequest(payload []byte) (*PTYRequest, error) {
 	req.HeightPx = binary.BigEndian.Uint32(payload[offset+12 : offset+16])
 
 	offset += 16
-	if offset < len(payload) {
+	if int(offset) < len(payload) {
 		modesLen := binary.BigEndian.Uint32(payload[offset : offset+4])
 		offset += 4
-		if offset+int(modesLen) <= len(payload) {
-			req.Modes = string(payload[offset : offset+int(modesLen)])
+		if int(offset)+int(modesLen) <= len(payload) {
+			req.Modes = string(payload[offset : uint32(int(offset)+int(modesLen))])
 		}
 	}
